+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - `npm run start-dev` => `npm run dev`
 - renamed, removed, and revamped other npm scripts
 - the nature of the dev flow has changed substantially (client-side assets are now generated via `webpack` embedded into the server-side module)
-- on production, everything should behave like before (no `webpack`, no `react-refresh`)
+- on production, everything should behave like before (no `webpack-dev-middleware`, no `react-refresh`)
+- the template `index.html` now has its `title` injected by `webpack` using the build-time env var `APP_NAME`
 
 ### Added
 - the use of `react-refresh` (via `webpack-dev-middleware` and `webpack-hot-middleware`) to support live-reload